+++ src/modules/Northbound/Adapter/OPCUAAdapter/index.ts
@@ -116,8 +116,8 @@ export class OPCUAAdapter {
           encoding: 'utf-8'
         });
         let xmlFileReadMachineNameFixed = xmlFileRead.replace(
-          new RegExp('{{machineName}}', 'g'),
-          await this.getMachineName()
+          new RegExp('DummyMachineToolName', 'g'),
+          await this.getHostname()
         );
 
         const xmlFileReadParsed = create(xmlFileReadMachineNameFixed);
@@ -157,7 +157,7 @@ export class OPCUAAdapter {
               continue;
             }
             const nodeId = `ns=1;s=${customConfig.nodeId}`;
-            const browseName = `1:${customConfig.nodeId}`;
+            const browseName = `2:${customConfig.nodeId}`;
             const name = customConfig.name;
             const dataType = customConfig.dataType;
 
@@ -401,9 +401,12 @@ export class OPCUAAdapter {
       });
   }
 
-  public findNode(nodeIdentifier: NodeIdLike): BaseNode | null {
+  public async findNode(nodeIdentifier: NodeIdLike): Promise<BaseNode | null> {
     const logPrefix = `${OPCUAAdapter.className}::findNode`;
-    const node = this.server.engine.addressSpace.findNode(nodeIdentifier);
+    const nodeIdentifierWithHostname = `ns=7;s=${await this.getHostname()}.${nodeIdentifier}`;
+    const node = this.server.engine.addressSpace.findNode(
+      nodeIdentifierWithHostname
+    );
 
     if (node) return node;
     winston.warn(`${logPrefix} Node with id ${nodeIdentifier} not found!`);