+++ user-docs/docs/CHANGELOG.md
@@ -3,11 +3,14 @@ title: Changelog
 ---
 
 # Changelog
+
 ## [2.2.2]
 
 ### Added
 
 - [DIGMDCLGHT-129](https://jira.app.dmgmori.com/browse/DIGMDCLGHT-129) Added support for custom data points.
+- [DIGMDCLGHT-175](https://jira.app.dmgmori.com/browse/DIGMDCLGHT-175) Added support for OPC UA schema version 2.0.18. Supports usage of config file between different machines. Ignores if custom given node id matches existing one.
+
 ## [2.2.1]
 
 ### Added