+++ user-docs/docs/CHANGELOG.md
@@ -4,6 +4,12 @@ title: Changelog
 
 # Changelog
 
+## [2.2.2]
+
+### Added
+
+- [DIGMDCLGHT-103](https://jira.app.dmgmori.com/browse/DIGMDCLGHT-103) Added support for custom mathematical expressions for virtual data points
+
 ## [2.2.1]
 
 ### Added